eeprom: support any 24Cx EEPROM configuration
Add `mode' argument to `eeprom_init' to specify the kind of
EEPROM to use.  This is then searched in a table with
corresponding EEPROM size, default page size (can be overridden
with _set_page_size) and offset (data address) size (1 or 2
bytes).  The implementation now has a dynamic sizes for the data,
page, and offset.

Tested with Microchip 24LC01 "24c01" and 24LC256 "24c256" only.